Fix for cublas

+++ src/cublas/cublas_create_allocate.cpp
@@ -77,7 +77,7 @@ long long get_malloc_size_host(mblas_data_type type, long x, long y, int batch, 
 
 long long get_malloc_size_dev(mblas_data_type type, long x, long y, int batch, long long stride) {
   int typesize = type_call_dev<sizeofCUDT>(type);
-  long packing_count = type.get_packing_count();
+  long long packing_count = type.get_packing_count();
   long long base = x * y;
   long long total_elements;
   if (batch > 1 && stride > base) {
